requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pandas==2.2.3
psycopg2-binary==2.9.10
supabase==2.9.1
transformers==4.46.0
torch==2.5.0
sentence-transformers==3.2.1
vaderSentiment==3.3.2
textblob==0.18.0.post0
gensim==4.3.3
matplotlib==3.9.2
seaborn==0.13.2
scikit-learn==1.5.2
python-dotenv==1.0.1
nltk==3.9.1
emoji==2.14.1
transformers==4.46.0
torch==2.5.0
langdetect==1.0.9
sentencepiece==0.2.0
uvicorn==0.32.0
sacremoses==0.1.1